put back actionSelectorCheck in control.cpp
