+++ backends/bmv2/Makefile.am
@@ -37,6 +37,7 @@ p4c_bm2_ss_UNIFIED = \
 	backends/bmv2/midend.cpp \
 	backends/bmv2/metermap.cpp \
 	backends/bmv2/parser.cpp \
+	backends/bmv2/sharedActionSelectorCheck.cpp \
 	backends/bmv2/v1model.cpp
 
 noinst_HEADERS += \
@@ -55,7 +56,8 @@ noinst_HEADERS += \
 	backends/bmv2/lower.h \
 	backends/bmv2/midend.h \
 	backends/bmv2/metermap.h \
-	backends/bmv2/parser.h
+	backends/bmv2/parser.h \
+	backends/bmv2/sharedActionSelectorCheck.h
 
 cpplint_FILES += $(p4c_bm2_ss_UNIFIED) $(p4c_bm2_ss_NONUNIFIED)
