+++ backends/bmv2/metermap.h
@@ -21,9 +21,6 @@ limitations under the License.
 
 namespace BMV2 {
 
-/**
- *
- */
 class DirectMeterMap final {
  public:
     struct DirectMeterInfo {
